Fix E365 ("File already opened in another process") prompts so they don't hang
Now that we're using nvim-rs for everything, being able to handle some of
the neovim requests that can potentially block infinitely on user input is
significantly simpler than before. Previously, neovim-gtk would assume that
executing commands for opening files would always be a non-blocking
operation. This isn't true however, as if the file is already being used by
a different process we'll get an error prompt asking us what action we'd
like to take - while simultaneously blocking the command request for
opening the file until the point at which the user has provided us input
which could potentially lead to the request timing out in neovim-lib.

Another issue with how we handled this in neovim-gtk is that if the user
does answer the prompt with certain options - the command request will
actually fail with a not unexpected call error.

So now that we're using nvim-rs, let's fix this by switching away from
spawn_timeout!() and instead issuing the aforementioned requests with no
timeout. We handle the request results asynchronously once the request has
completed, and automatically quiet reporting of any expected call errors.
We also add a is_blocked() helper function to NvimSession, which we make
use of in order to avoid attempting to send FocusIn/FocusOut requests to
neovim while neovim is blocking on input from the user for E365 prompts (as
such requests will also block and potentially timeout when neovim is
waiting input).

This mostly fixes the issue, and prevents neovim-gtk from hanging when a
file is already in use by another process. Note however that we still need
to deal with the issue of the async requests from various GUI elements (the
new tab button, Save All button, etc.) to prevent them from being sent and
potentially timing out while neovim is awaiting user input. We could simply
do this part with our new NvimSession.is_blocked() helper, but a nicer
solution I'd like to implement would be dynamically enabling/disabling GUI
buttons before/after beginning such neovim requests.