Send resize requests immediately instead of intervallically
Queue up resize requests immediately instead of only once every 200ms,
keeping only the most recent unsubmitted resize in the queue at a time.
Additionally since resize requests can block indefinitely, we do this
without using a timeout.

The end result is that resizing becomes a lot more responsive, and doesn't
result in timeout errors when neovim is blocking. The only downside is that
when we eventually add support for remote neovim sessions, it's possible we
might want to add back the ability to optionally send requests
intervallically but with async code instead of using glib timeouts.