Disable UI elements when we know we're about to block on a prompt
This doesn't yet fully solve the issue of being able to use UI elements
while nvim is blocked, but it does fix it in cases where we know for sure
a request we're about to send has a chance of blocking. We also make it so
that we start nvim-gtk with header bar elements disabled by default until
they're ready to be used.