Add support for executing VimL commands from cmdline after opening files
This is analogous to neovim's -c. We implement this because due to the fact
that both the commands for opening files on start and commands passed
directly to neovim via -- using the -c option can potentially block and
cause neovim-gtk to hang. If we handle sending the commands ourselves
though, we can send them along with our commands for opening the first few
files - allowing us to nicely disable GUI elements while the commands are
running, and handle anything that might potentially block us.