Don't update Pmenu, PmenuSel, or Cursor when receiving ui highlight items
Ever since switching to nvim-cmp for my autocompletion at work, I've been
noticing that the background color of the native popup menu will
occasionally change background colors when displaying hover documentation.

After doing a bunch of debugging, I eventually traced the change in
background colors down to `HighlightMap::set()`. What appeared to be
happening was that we would receive a `hl_attr_define` event with info
containing multiple semantic highlighting items like so:
be something like this:

  ```
  kind: ui
  ui_name: NormalFloat
  hl_name: Pmenu
  …

  kind: syntax
  hl_name: PreProc
  …
  ```

If I understand this excerpt of documentation from nvim v0.5.1 correctly:

  > Note: "ui" items will have both `ui_name` and `hi_name` present. These
  > can differ, because the builtin group was linked to another group
  > `:hi-link`, or because 'winhighlight' was used. UI items will be
  > transmitted, even if the highlight group is cleared, so `ui_name` can
  > always be used to reliably identify screen elements, even if no
  > attributes have been applied.

This likely means that we're not supposed to interpret UI events as syntax
item changes. And as it turns out, not interpreting them as syntax item
changes totally fixes the popup menu background coloring. Hooray!