Handle multiple items sharing a cell in the renderer (V2)
This is a somewhat nasty bug I found by accident and have been stuck on for
a while. Basically, a string like this:

  Fö̲o

Would break rendering due to the fact that in certain fonts, the combining
characters in this string might require the use of other fonts in order to
be rendered - resulting in pango's itemizer splitting a cell into multiple
items instead of just one.

Because we currently only assume one Item per Cell, this causes all sorts
of problems including crashes and breaking the rendering of any glyphs
coming after a combining character.

So, let's fix this by storing multiple Items per Cell and rendering them
one at a time onto each cell when redrawing. This still won't result in
totally correct glyph rendering, which we'll address in the next commit.
But it will stop us from crashing, and allow us to display something
vaguely resembling the correct result and render the rest of the line.

V2: Undo the changes that I did with trying to get rid of having to clone
pango items, that seems to be what broke things.