Switch fonts when favorable for pango itemizer results (V2)
Occasionally when rendering text it's possible that we can run into
characters which require glyphs that aren't provided by the current fonts
that we're using. When this happens, pango's itemizer is forced to split
glyphs with different fonts into separate items. This can cause problems
however, if we're trying to render combining characters and the combining
characters in question end up in a separate item than the character they're
meant to combine with. The reason being that we may end up getting a
different glyph that would typically only be used for the combining
character in the absence of a character to combine with - resulting in
incorrect glyph output.

So, fix this by attempting to retry itemization on items that get split up
while trying the various different fonts glyphs were taken from to find the
one that gives us as few items as possible - giving us a much better chance
of actually rendering the correct glyphs.

This isn't perfect unfortunately, but should be a pretty big improvement.

V2: Undo the changes that I did with trying to get rid of having to clone
pango items, that seems to be what broke things.