Fix undercurl height calculation
The code in draw_underline_strikethrough that's used for determining the
height to draw undercurls is rather bizarre. As far as I can tell, it
appears to try to somehow estimate the size of the undercurl depending on
where an underline would be typically drawn - and then multiplies this +
the underline width against some seemingly arbitrary looking values.

Regardless of whatever the original idea was, this doesn't end up actually
working all of the time. Instead, this occasionally results in very broken
looking undercurls on fonts such as Inconsolata Nerd Font (see issue
mentioned below for a link).

Lucky for us however, there's actually a much simpler way we can determine
the height of the undercurl. Since we draw the undercurl by providing the
coordinates of a rectangle in which the undercurl is drawn aligned to the
vertical center of the rectangle - we can just compute the height and y
coordinates of this box by assuming it starts at the beginning of the
descent and height equivalent to the descent.

So, pipe through the descent value for each cell and calculate the
undercurl height that way.

Fixes #11