Fix underline coloring in the absence of a special color
Ever since fixing #3, we've had the issue of neovim-gtk coloring underlines
as red in many situations. The reason for this being that when neovim sends
a set of default colors to be used, it comes up with a default color on
it's own in the absence of an explicit value in a given highlighting rule.
Unfortunately neovim seems to choose the color red for special, which
results in underlines defaulting to red unless a different color is
specified.

Ideally we should be making it so that only undercurls default to red, and
make underlines default to the foreground color when no color is specified
for a highlighting group. Luckily with the addition of the ext_termcolors
UI capability to neovim, we can have neovim tell us when there's no color
set for a highlighting group instead of it giving us a default value! So,
let's take advantage of this and use this in order to come up with correct
underline colors. For normal background and foreground colors, we'll try to
follow the same logic that neovim describes in it's documentation for
providing default colors based on the value of the 'background' setting.

This does have the unfortunate requirement of us needing to temporarily
switch over to my fork of nvim-rs in order to implement support for the
ext_termcolors capability. I've already begun the process of upstreaming
these changes at https://github.com/KillTheMule/nvim-rs/pull/39 , and will
remove the dependency on this fork once this is merged.

Fixes #10