Properly handle :cq
When the user invokes :cq, this causes Neovim to exit unsuccessfully.
This is important if the user wishes to abort a Git commit message, or
in other situations where the operation is aborted when the editor exits
unsuccessfully.

Let's make this work by listening to the VimLeavePre autocommand and
then reading the v:exiting code, which indicates the exit status with
which Neovim is exiting.  Use an ApplicationCommandLine object to store
this exit status, and propagate it from the main program, where we get
it, into the shell's state, where we can modify it.  We cannot pass this
into the ShellOptions structure, as much as we would like to, because
it results in that structure not being Send, which causes problems
elsewhere.

Note that this object does not intrinsically modify the exit status of
the program, nor does the callback from which we get it, despite what
the documentation says.  We must do so manually at the end of the
program for this to work properly.

Because we must set the HANDLES_COMMAND_LINE flag in order to get this
callback, we can no longer use the open or activate callbacks or
run_with_args.  Change to handle the argument parsing ourselves in this
command_line callback so that we get the same behavior as before, and
change the various required things to make our arguments line up with
the new callers.