Open/close "projects" popup by activating/deactivating MenuButton
Instead of calling popup()/popdown() directly and using the clicked signal,
simply use the MenuButton's built-in popover handling - which allows the
"Open" button's relief to stay in sync with whether the project view is
visible or not.

Additionally for setting the correct size on the tree view in the popup, we
now rely on the 'map' event since we're no longer handling the click events
directly.