Don't use predictably named temporary files
When neovim-gtk starts and it daemonizes, it writes two temporary files
into the /tmp directory based on the name of the user.  Unfortunately,
this is insecure, since an attacker could create those files before us
as symlinks pointing to a file they'd like us to write junk to, and then
we would do exactly that.  The only secure way to create a temporary
file is by using O_EXCL, and unix_daemonize doesn't do that.

On some Linux distros, this is prevented by fs.protected_symlinks and
friends, but in general, we can't safely assume that the user has this
flag enabled, and this wouldn't be true of other OSes where neovim-gtk
might run, so let's just switch to redirecting to /dev/null.

We could do something like redirect to .xsession-errors, but that might
not be present on Wayland. If someone really wants to see these
messages, they can do so by running with --no-fork, but they probably
don't care and just discarding them is fine.

Since we no longer need the whoami crate, remove it from Cargo.toml and
Cargo.lock.