Add NVIM_GTK_CMD_AS_META env var option to treat Command as Alt on macOS
No modifier states are set in GTK when holding the Option key, but they
are when holding the Command key. Since that key is also unused in nvim
itself, this can be used as a replacement for the Alt/Meta key on other
keyboards.

However, because that key is often used for system commands, like cmd+v
to paste, this behavior can be selectively enabled by setting the
`NVIM_GTK_CMD_AS_META` environment variable, similarly to the existing
`NVIM_GTK_` environment variables used for enabling/disabling UI features.

Fixes #23

(Lyude squashed the commits from this PR and modified the commit message)