Introduce NvimViewport
Since GTK4 doesn't have a concept of damage when it comes to redraws (e.g.
we redraw the whole frame at once), we want to introduce a new widget that
takes advantage of GTK4's new snapshot rendering infrastructure. This will
provide us something that's partially hardware accelerated, and doesn't
really need damage calculation!

We'll call this NvimViewport, and we'll port all of the cairo
infrastructure to rendering over to using snapshots. We don't drop
DrawingArea quite yet, since the cmdline widget is still relying on it for
now. It should be noted this widget also isn't exactly the most efficient
snapshot implementation we could get - but let's focus on getting things
working first and optimize later.