ui: Set background via custom CSS class
The built in 'background' class was removed from the GtkWindow, causing
regression, specifically the titlebar having the wrong foreground color
when not using the dark theme.

The class was removed to make transparency work, but to make it work,
without removing builtin CSS classes, just create our own class and
override the background with rgba(0,0,0,0).