popup_menu: Get rid of popupmenu_show/hide() hacks, use flushes instead
While moving popup menus over to GtkListView I noticed that we drop various
popupmenu_show/popupmenu_hide() events that come in from neovim. It appears
the reason this was done previously was because we didn't do any actual
handling for the flush event, we needed another way to avoid processing
intermediate popupmenu_hide events that might come before the end of a
"batch" of events.

I'm not 100% sure on this change, but getting rid of this hack definitely
makes things substantially more responsive. As well in order to ensure we
don't hide the popupmenu on an intermediate hide event, we queue up
popupmenu actions in a similar manner to how we handle redraw events. So I
think this should be safe enough to go with.