popup_menu: Fix double borrow_mut() error properly this time
It turns out that simply firing off the popup menu in an idle callback
isn't enough to fix this problem. It seems what's likely actually happening
here is a fix in GTK 4.6 regarding behavior with multiple application
windows that triggered a nvim-gtk bug.

Since the popup menu is technically another window, moving the cursor above
it results in the cursor traveling away from the main application window
over to the popup menu window. Moving between windows with the cursor
usually triggers an input event on the respective surfaces, but it appears
GTK 4.4 would make the mistake of not sending this event when the popup
menu surface was hidden. GTK 4.6 fixes this, but this triggers a bug in
nvim-gtk.

Currently our motion events in nvim-gtk borrow the state mutably, and since
we have portions of the popup menu in the global shell state - this results
in us attempting to borrow the state as mutable twice leading to a crash.
Luckily for us, looking closely enough at the motion event code reveals
that there's not actually a need for us to have a mutable reference to the
shell state in those functions - only an immutable one.

So, in order to fix this we make the motion events only grab a immutable
reference to the global shell state. We then move pretty much all of the
popup menu state in the PopupMenuState structure, which has it's own
RefCell and can be borrowed immutably while only requiring that we grab the
shell state immutably. Since we can have infinite immutable references to a
RefCell at once, this fixes the issue without a need for an idle loop.