render: Fix undercurls in GTK 4.6+
Turns out that the code I wrote for undercurls looks beautiful, but also
breaks on GTK 4.6 despite working on 4.4. My assumption is this isn't a bug
on GTK+'s side, but likely some reasonable change to how rendering is
handled that changed some implicit behavior we shouldn't have been
previously relying on.

So, let's fix this. This looks just a little bit uglier, but getting
undercurls looking perfect with snapshots is surprisingly difficult.

Also, I think this is the last regression I've hit on GTK4 :)