render: Handle text node generation failures gracefully
As made clear by the testcase here:

  https://github.com/Lyude/neovim-gtk/issues/8#issuecomment-1353840913

It's totally possible for us to get screens of garbage content that lead to
situations where we just fail to generate a TextNode. Unfortunately I don't
really see much of an easier way to handle this (perhaps these are control
characters that we need to generate boxed hex-code glyphs for?) for the
time being, so let's just handle the possibility of no TextNode.