render: Align start of repeating segment for undercurls to its own width
Turns out that the method I came up with for drawing undercurls in GTK+
>4.6 wasn't totally correct. Because we were aligning the start of the
repeating segment to the diameter of the dots in the undercurl, this made
it so that we could potentially begin the first dot of the undercurl
precisely where an undercurl ended on the last cell. This also made it so
the dots would occasionally move if you put the text cursor over them.

So, let's fix that by aligning the starting X coordinate to a multiple of
the width of the entire repeating segment (including the empty space coming
after the dot).

Also, we start specifying the child bounds for the repeating segment here
instead of enlarging the segment by drawing a transparent square at the
end.