nvim: Begin processing metadata in nvim_get_api_info() with supported ui options
We'll start by hooking up support for extension checks at startup, so we
can tell which ones we can use ahead of time - and throw an error if one we
really need isn't present.

I -think- that we need linegrid to do anything useful, and that it is a
somewhat recent extension, but I'm not entirely sure as I haven't bothered
checking. Hopefully no one bisects this commit!