popup_menu: Add support for nvim_ui_pum_set_bounds()
This will allow us to report the current bounds of the native popup menu to
neovim, which allows neovim to adjust the position of additional windows so
they're not hidden under the popup menu. In order to do this, we introduce
a new widget called PopupMenuPopover - which is just a popover, but with a
signal we can hook up to in order to get notified of position changes.