Bump version to 1.1.0-devel
This isn't an actual version release, but a version bump so that we can
start maintaining a separate stable v1.0 branch, which only contains fixes
instead of new features. We'll mark non-stable releases as -devel.

Ideally things shouldn't be too unstable on the main branch though, since I
use the main branch as a daily driver :).