Improve cursor color defaults
When deciding which cursor color to use in the absence of a bg/fg color, we
make the mistake of using the color as-is. This means in the case of
setting the cursor color for a setup without any `*fg` or `*bg` values, we
make the cursor match the background color - which makes it invisible.
Whoops.

What we really want here is to use the closest thing we can to the inverse
of the default background color. Luckily we forgot to actually drop the
`inverse()` function that we were using before `fade()` during the GTK4
conversion, so let's convert this to using a fixed inverse_value of 1.0 and
use this to determine the default cursor bg. This should provide something
that looks more legible in most setups hopefully