cursor: Replace Cursor trait with BlinkCursor
For whatever reason, daa84 used to have two structs which would implement a
main Cursor trait to represent different states the cursor could be in.
This included a BlinkCursor, and an EmptyCursor.

Somewhere down the line, it got kind of painful to do this and made a lot
more sense to just represent this with Option<Cursor> and drop EmptyCursor.
This just left us with BlinkCursor, the only implementor of Cursor.

Since this is kind of ugly and we don't need this anymore: drop the Cursor
trait and move it's contents into BlinkCursor, then rename BlinkCursor to
Cursor.