cursor: Don't show intermediate cursor positions to the user
While the rest of the ui_model is updated atomically through virtue of the
fact we usually have a cached Snapshot to use in-between screen updates:
the cursor is the only thing that isn't. This is mainly because we need to
re-render the cursor position more often then the screen updates, in order
to have a pulsing effect.