render: Don't use Option<…> for atomic cursor updates
It occurred to me it's actually less error prone in the long run if we
don't use an option here, and instead refactor this so it's clearer that
you usually want the "real" cursor position as opposed to the flushed
position.

We also fix UiModel::cur_point(), which was a victim of getting set to use
the wrong cursor position by mistake.