ui: Implement Drop for UiMutex<…>
I was looking at the `glib::ThreadGuard` struct and wondering why it didn't
seem to provide `Sync` and only `Send`. This caused me to look at our own
ui::UiMutex and figure out how it worked. Unfortunately now that I've
stared at it for a while, it's occurred to me there's likely nothing
_actually_ guaranteeing the effects of `Sync` here. The reason I think this
is because an object which is `Send` + `Sync` would at a bare minimum
require a destructor which could atomically dereference the data and
destroy it in a thread-safe manner. `UiMutex` technically provides a
read/write barrier, but not in the context of destruction.

This isn't to say `UiMutex` in it's current form is "broken" per-say,
though. In just about every instance I can see, we wrap `UiMutex` in an
`Arc` which should fulfill the aforementioned requirement. This does mean
for us however, that it's technically possible that the last reference to
an `Arc<UiMutex>` could get dropped on a thread other than the one which
the `UiMutex` was created on - which would be undefined behavior! Big
yikes.

Luckily as far as I can tell, this should never happen so long as we always
make sure to shut down the tokio runtime (and any native threads we've
spawned) before GTK+ shuts down - which should get rid of the chance of us
having any such thing happen. It's less then ideal to be allowed by the
compiler to write code like this, but only having to deal with a single
place where this could go wrong if we're not careful isn't too bad.

This being said though - currently we aren't actually doing any runtime
assertions to verify this is the case, which makes me quite uneasy! So,
let's copy how `glib::ThreadGuard` handles this and add a runtime assertion
for this in our `Drop` implementation to _really_ make sure this never ever
happens.

(turns out after I wrote this and tried quitting neovim-gtk, this found a
single bug! Hooray for my intuition from C :)