Drop code for delayed redraw events
We're about to turn on runtime-drop checks for UiMutex to make it
_actually_ hold up to it's promise of being `Send` and `Sync` (which it
currently does not do) to ensure we only ever drop `UiMutex` objects on the
thread they were created. Turns out we currently aren't doing this for the
delayed redraw event we store here - which will cause this code to panic.

We don't even use this code anymore, so just drop all of it.